+++ src/components/shared/forms/RenameWorkspaceForm.tsx
@@ -4,7 +4,7 @@ import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useDispatch, useSelector } from 'react-redux'
 import { AppDispatch, RootState } from '../../../redux/store'
-import { workspaceShema } from '../../../schema/workspace'
+import { workspaceShema } from '../../../schema/modal'
 import { getWorkspacesByUser, renameWorkspace } from '@/redux/slice/workspaceSlice'
 import toast from 'react-hot-toast'
 import { AxiosError } from 'axios'
