+++ src/pages/ManageWorkspaces.tsx
@@ -8,7 +8,7 @@ import { createWorkspace, getWorkspacesByUser } from '../redux/slice/workspaceSl
 import DialogDemo from '@/components/shared/shared/Modal'
 import WorkspaceForm from '@/components/shared/forms/WorkspaceForm'
 import { z } from 'zod'
-import { workspaceShema } from '@/schema/workspace'
+import { workspaceShema } from '@/schema/modal'
 import toast from 'react-hot-toast'
 import { handleAxiosError } from '@/util/helpers'
 import { AxiosError } from 'axios'
